requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.2

# Utilities
tqdm>=4.65.0
